+++ pages/contact.js
@@ -2,7 +2,8 @@ import { useState } from 'react';
 import ContactUsFormSubscribe from '@/components/ContactUs';
 import ContactUsCards from '@/components/ContactUs/ContactUsCards';
 import S from '@/styles/pages/contactStyles';
-
+import Bracket from '@/components/decorations/Bracket';
+import bracketStyles from '@/components/decorations/Bracket/Bracket.module.scss';
 export default function ContactUs() {
   const [message, setMessage] = useState([]);
 
@@ -10,7 +11,7 @@ export default function ContactUs() {
     <>
       <S.ContactUsContainer>
         <S.FormAndDecorations>
-          <S.YellowBracket />
+          <Bracket className={bracketStyles.yellowBracket} />
           <ContactUsFormSubscribe setMsg={setMessage} />
           <S.YellowColon src='/images/svg/yellow-colon.svg' />
           <S.ResponseMessage>