+++ styles/pages/contactStyles.js
@@ -1,5 +1,4 @@
 import styled, { css } from 'styled-components';
-import Bracket from '@/components/decorations/Bracket';
 
 const ContactUsContainer = styled.div`
   background-color: ${({ theme }) => theme.colors.lightBg};
@@ -43,40 +42,6 @@ const FormAndDecorations = styled.div`
   `}
 `;
 
-const YellowBracket = styled(Bracket)`
-  -webkit-transform: scaleX(-1);
-  transform: scaleX(-1);
-
-  ${props => css`
-    @media (max-width: ${props.theme.breakpoints.tablet}) {
-      display: none;
-    }
-  `}
-
-  ${props => css`
-    @media (min-width: ${props.theme.breakpoints.tablet}) {
-      position: absolute;
-      top: 28%;
-      right: 120%;
-      width: 8rem;
-    }
-  `}
-
-  ${props => css`
-    @media (min-width: ${props.theme.breakpoints.desktop}) {
-      top: 25%;
-      right: 120%;
-    }
-  `}
-
-  ${props => css`
-    @media (min-width: ${props.theme.breakpoints.desktopPlus}) {
-      top: 45%;
-      right: 140%;
-    }
-  `}
-`;
-
 const YellowColon = styled.img`
   ${props => css`
     @media (max-width: ${props.theme.breakpoints.desktopMinus}) {
@@ -115,7 +80,6 @@ const ResponseMessage = styled.div`
 export default {
   ContactUsContainer,
   FormAndDecorations,
-  YellowBracket,
   YellowColon,
   ResponseMessage,
 };